"Changed target dependencies to triggers in build files"
Modified the dependencies of various targets in `Build.Compile.cs`, `Build.Installer.cs`, and `Build.Sign.cs` to triggers. This means that the dependent target will now be executed if the target it depends on is executed, but not vice versa. Specifically, the `Compile` target is now triggered by the `Clean` target, the `Installer` target by the `Sign` target, and the `Sign` target by the `Compile` target. Also, the main method in `Build.cs` has been updated to execute the `Clean` target by default instead of the `Installer` target.